requirements.txt
# Florida County RAG - Dependencies
# Install: pip install -r requirements.txt

# PDF Processing
pymupdf>=1.23.0

# Vector Database
chromadb>=0.4.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0

# Web UI
streamlit>=1.30.0

# For development/testing (optional)
# pytest>=7.4.0
# black>=23.0.0
# ruff>=0.1.0
